[FEM] Elmer writer: sort out electricforce equation
- sort out electricforce equation - next step to refactor writer.py
