+++ src/Mod/Fem/CMakeLists.txt
@@ -253,6 +253,7 @@ SET(FemSolverElmerEquations_SRCS
     femsolver/elmer/equations/__init__.py
     femsolver/elmer/equations/elasticity.py
     femsolver/elmer/equations/electricforce.py
+    femsolver/elmer/equations/electricforce_writer.py
     femsolver/elmer/equations/electrostatic.py
     femsolver/elmer/equations/electrostatic_writer.py
     femsolver/elmer/equations/equation.py
