+++ src/Mod/Fem/femsolver/elmer/writer.py
@@ -49,7 +49,7 @@ from femtools import constants
 from femtools import femutils
 from femtools import membertools
 from .equations import elasticity
-from .equations import electricforce
+from .equations import electricforce_writer as EF_writer
 from .equations import electrostatic_writer as ES_writer
 from .equations import flow_writer
 from .equations import flux_writer
@@ -854,6 +854,7 @@ class Writer(object):
     # Electricforce
 
     def _handleElectricforce(self):
+        EFW = EF_writer.EFwriter(self, self.solver)
         activeIn = []
         for equation in self.solver.Group:
             if femutils.is_of_type(equation, "Fem::EquationElmerElectricforce"):
@@ -861,35 +862,9 @@ class Writer(object):
                     activeIn = equation.References[0][1]
                 else:
                     activeIn = self.getAllBodies()
-                solverSection = self._getElectricforceSolver(equation)
+                solverSection = EFW.getElectricforceSolver(equation)
                 for body in activeIn:
                     self._addSolver(body, solverSection)
-
-    def _getElectricforceSolver(self, equation):
-        # check if we need to update the equation
-        self._updateElectricforceSolver(equation)
-        # output the equation parameters
-        s = self._createEmptySolver()
-        s["Equation"] = "Electric Force"  # equation.Name
-        s["Procedure"] = sifio.FileAttr("ElectricForce/StatElecForce")
-        s["Exec Solver"] = equation.ExecSolver
-        s["Stabilize"] = equation.Stabilize
-        return s
-
-    def _updateElectricforceSolver(self, equation):
-        # updates older Electricforce equations
-        if not hasattr(equation, "ExecSolver"):
-            equation.addProperty(
-                "App::PropertyEnumeration",
-                "ExecSolver",
-                "Electric Force",
-                (
-                    "That solver is only executed after solution converged\n"
-                    "To execute always, change to 'Always'"
-                )
-            )
-            equation.ExecSolver = electricforce.SOLVER_EXEC_METHODS
-            equation.ExecSolver = "After Timestep"
 
     #-------------------------------------------------------------------------------------------
     # Flow
@@ -961,7 +936,7 @@ class Writer(object):
     #-------------------------------------------------------------------------------------------
     # Solver handling
 
-    def _createEmptySolver(self):
+    def createEmptySolver(self):
         s = sifio.createSection(sifio.SOLVER)
         return s
